requirements.txt
fastapi
uvicorn
SQLAlchemy
 pydantic
requests
PyMySQL
Jinja2
python-docx
